[py_connector] split the vLLM connector into per-role cores
The connector was one 1348-line class serving both roles through if-role
branches: scheduler hooks (matching, save orchestration, finishing) and
worker hooks (translation, data-plane transfer) interleaved on shared
state, and every reader had to know which _alive_requests / counter pair
belonged to which role's instance.

Composition over inheritance (the shared surface is data classes and pure
functions, not behavior -- no base class):

* vllm_common.py -- role-agnostic vocabulary: GroupMeta / ReqState,
  spec naming + spec-group building, KV layout normalization, the hybrid
  capability gate, parse_groups. May not import the cores.
* scheduler_core.py -- SchedulerCore: the authoritative _alive_requests,
  external matching (_safe_external_prefix, _external_match_burned),
  per-step metadata assembly, save orchestration, request finishing.
* worker_core.py -- WorkerCore: transfer client setup, TransferGroup
  building, block index translation, load/save task planning, the
  mirrored _alive_requests replayed from scheduler deltas. Per-step
  metadata is passed in explicitly -- the core holds no metadata state.
* v1_connector.py -- 234-line shell: role-agnostic setup (config,
  registration) plus one-line hook delegation. vLLM instantiates one
  class per role from the plugin registry, so the shell stays the entry
  point. Compatibility re-exports keep external import paths (tests,
  e2e harness) stable.

Method bodies moved verbatim; the only semantic change is the explicit
meta parameter on WorkerCore.start_load_kv / wait_for_save /
get_finished (previously read through the vLLM base-class field).

Tests: factories now build the core under test (make_connector ->
WorkerCore for translation, new make_scheduler_core for the scheduler
loop); TestParseGroups calls the module function; the e2e
MutatedConnector swaps in a MutatedWorkerCore subclass -- overriding the
shell method no longer reaches the translation after the split. 102
connector cases green; real-vLLM (0.26) import chain verified.
